Moves serialize to core
