+++ packages/unmock-core/src/serialize/interfaces.ts
